rewrite to use [add|remove]ObserverForEvent
Primarily so the object will only observe events for which there are triggers (instead of observing all of them all the time). This also allowed some simplification and removal of unwieldy if-else chains.

It would be much simpler to just use the actual notification names instead of having to maintain maps, but that would break users' existing event triggers.